order_executor/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
watchdog==6.0.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0